requirements.txt
anthropic>=0.18.1
python-dotenv>=1.0.0
Flask>=3.0.0
gunicorn>=21.2.0 